Rename/refactor based on latest concepts
