+++ langkit/core/demo.py
@@ -5,10 +5,10 @@ from typing import List, Mapping
 import pandas as pd
 
 from langkit.metrics import lib
-from langkit.metrics.metric import MetricCreator, MetricNameCapture, MetricResult, MultiMetric, MultiMetricResult
+from langkit.core.metric import MetricCreator, MetricNameCapture, MetricResult, MultiMetric, MultiMetricResult
 from langkit.metrics.regexes.regex_loader import CompiledPattern, CompiledPatternGroups
-from langkit.pipeline.pipeline import EvaluationWorkflow, Hook
-from langkit.pipeline.validation import ValidationResult, create_validator
+from langkit.core.workflow import EvaluationWorkflow, Hook
+from langkit.core.validation import ValidationResult, create_validator
 
 pd.set_option("display.max_columns", None)
 pd.set_option("display.width", None)
