+++ langkit/metrics/sentiment_polarity.py
@@ -4,7 +4,7 @@ import nltk
 import pandas as pd
 from nltk.sentiment import SentimentIntensityAnalyzer
 
-from langkit.metrics.metric import Metric, SingleMetric, SingleMetricResult, UdfInput
+from langkit.core.metric import Metric, SingleMetric, SingleMetricResult, UdfInput
 
 
 def __sentiment_polarity_module(column_name: str, lexicon: str = "vader_lexicon") -> Metric:
