+++ tests/langkit/metrics/test_sentiment_polarity.py
@@ -4,7 +4,7 @@ from typing import Any
 import pandas as pd
 
 import whylogs as why
-from langkit.metrics.metric import EvaluationConfifBuilder, EvaluationConfig
+from langkit.core.metric import EvaluationConfifBuilder, EvaluationConfig
 from langkit.metrics.sentiment_polarity import promp_response_sentiment_polarity, promp_sentiment_polarity, response_sentiment_polarity
 from langkit.metrics.whylogs_compat import create_whylogs_udf_schema
