+++ tests/langkit/metrics/test_toxicity.py
@@ -4,7 +4,7 @@ from typing import Any
 import pandas as pd
 
 import whylogs as why
-from langkit.metrics.metric import EvaluationConfifBuilder, EvaluationConfig
+from langkit.core.metric import EvaluationConfifBuilder, EvaluationConfig
 from langkit.metrics.toxicity import prompt_response_toxicity_module, prompt_toxicity_module, response_toxicity_module
 from langkit.metrics.whylogs_compat import create_whylogs_udf_schema
